requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.9.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
dataclasses-json==0.6.7
debugpy==1.8.2
decorator==5.1.1
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
faiss-cpu==1.11.0
fastapi==0.115.12
filelock==3.18.0
frozenlist==1.4.1
fsspec==2025.3.2
greenlet==3.0.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
idna==3.7
interchange==2021.0.4
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.6
jiter==0.9.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.3.25
langchain-community==0.3.23
langchain-core==0.3.58
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.8
langsmith==0.3.42
Markdown==3.8
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib-inline==0.1.7
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
openai==1.77.0
openpyxl==3.1.5
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pansi==2024.11.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.2.2
prompt_toolkit==3.0.47
protobuf==6.30.2
psutil==6.0.0
pure-eval==0.2.2
py2neo==2021.2.4
pyahocorasick==2.1.0
pydantic==2.8.2
pydantic-settings==2.9.1
pydantic_core==2.20.1
Pygments==2.18.0
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
pywin32==306
PyYAML==6.0.1
pyzmq==26.0.3
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.46.2
sympy==1.14.0
tenacity==8.5.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.34.2
wcwidth==0.2.13
yarl==1.9.4
zstandard==0.23.0
